+++ lib/recorder.ps1
@@ -130,6 +130,15 @@ class Recorder : IRecorder {
             }
         }
     )
+    
+    hidden $_eject = $($this | Add-Member ScriptProperty 'eject' `
+        {
+            return Write-Warning ('ERROR: Recorder.eject is write only')
+        } `
+        {
+            $this.remote.Setter('Command.eject', 1)
+        }
+    )
 
     [void] Load ([string]$filename) {
         $this.Setter('load', $filename)